Update course action and datatable

+++ resources/views/backend/admin/course/action.blade.php
@@ -4,7 +4,7 @@
             <a href="{{ route('courses.edit', $course->id) }}" title="{{ __('edit') }}"><i class="las la-edit"></i></a>
         </li>
     @endif
-    @if(hasPermission('courses.destroy'))
+    @if(hasPermission('courses.destroy') && (isset($total_courses) ? $total_courses > 1 : true))
         <li>
             <a href="javascript:void(0)" onclick="delete_row('{{ route('courses.destroy', $course->id) }}', {{ $course->id }})" title="{{ __('delete') }}"><i class="las la-trash-alt text-danger"></i></a>
         </li>
@@ -34,7 +34,7 @@
                 <li><a class="dropdown-item"
                        href="{{ route('courses.edit',[$course->id,'tab'=>'faq']) }}">{{ __('faq') }}</a></li>
             @endif
-            @if(hasPermission('courses.destroy'))
+            @if(hasPermission('courses.destroy') && (isset($total_courses) ? $total_courses > 1 : true))
                 <li>
                     <a class="dropdown-item text-danger" href="javascript:void(0)" onclick="delete_row('{{ route('courses.destroy', $course->id) }}', {{ $course->id }})"><i class="las la-trash-alt me-2"></i>{{ __('delete') }}</a>
                 </li>
